LegalMind-AI/ai-engine/requirements.txt
fastapi
uvicorn[standard]
gunicorn
uvicorn
pydantic
langgraph
langchain
pinecone
sentence-transformers
PyMuPDF
requests
python-dotenv
langchain-groq
langchain-text-splitters
langchain-community
python-multipart
pinecone-text[splade]
nltk
ragas
datasets
pandas
langchain-huggingface
tabulate